+++ design-system/packages/button/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "@nathsimpson/button",
+  "author": "Nathan Simpson",
+  "description": "Button components",
+  "version": "0.1.0",
+  "main": "dist/nathsimpson-button.cjs.js",
+  "module": "dist/nathsimpson-button.esm.js",
+  "peerDependencies": {
+    "react": "^17.0.2"
+  },
+  "devDependencies": {
+    "react": "^17.0.2"
+  }
+}
