+++ design-system/packages/button/src/index.tsx
@@ -1,0 +1,1 @@
+export * from './Button';
