+++ website/src/components/Mdx.js
@@ -4,7 +4,7 @@ import { jsx } from '@emotion/core';
 import { MDXProvider } from '@mdx-js/react';
 import { MDXRenderer } from 'gatsby-plugin-mdx';
 
-import { Button, LinkButton, IconButton } from '@design-system/button';
+import { Button, LinkButton, IconButton } from '@nathsimpson/button';
 import { Icon } from '@nathsimpson/icon';
 import { Tiles } from '@design-system/tiles';
 import { Tag, Tags } from '@design-system/tag';
